fix: 순환 참조 해결을 위해 서비스 분리
